move firebase config settins to single file
